broke availabilty page to several components
