+++ client/src/App.tsx
@@ -30,7 +30,6 @@ function App(): JSX.Element {
                 <Route exact path="/signup" component={Signup} />
                 <Route exact path="/dashboard" component={Dashboard} />
                 <Route path="/profile/settings" component={Settings} />
-                <Route exact path="/profile/settings/Availability" component={Availability} />
                 <Route path="/profile/:id" component={ProfileDetail} />
 
                 <Route path="*">
