+++ client/src/pages/Settings/Availability/Availability.tsx
@@ -1,6 +1,5 @@
-import DropDownItem from './DropDownItem';
-function Availability(): JSX.Element {
-  return <DropDownItem />;
-}
+import Schedule from '../../../components/AvailabityTable/schedule';
+
+const Availability: React.FC = () => <Schedule />;
 
 export default Availability;
