+++ client/src/pages/Settings/EditProfile/EditProfile.tsx
@@ -99,11 +99,11 @@ const EditProfile: React.FC<EditProfileProps> = ({ header, currentUser, currentP
               id="name"
               label="Name"
               fullWidth
-              margin="normal"
+              //margin="normal"
               name="name"
               placeholder="Your name"
               autoComplete="name"
-              helperText={touched.name ? errors.name : ''}
+              helpertext={touched.name ? errors.name : ''}
               error={touched.name && Boolean(errors.name)}
               value={values.name}
               onChange={handleChange}
@@ -113,11 +113,11 @@ const EditProfile: React.FC<EditProfileProps> = ({ header, currentUser, currentP
               id="email"
               label="Email address"
               fullWidth
-              margin="normal"
+              //margin="normal"
               name="email"
               placeholder="Your email"
               autoComplete="email"
-              helperText={touched.email ? errors.email : ''}
+              t={touched.email ? errors.email : ''}
               error={touched.email && Boolean(errors.email)}
               value={values.email}
               onChange={handleChange}
@@ -129,7 +129,7 @@ const EditProfile: React.FC<EditProfileProps> = ({ header, currentUser, currentP
               margin="dense"
               name="gender"
               placeholder="Gender"
-              helperText={touched.gender ? errors.gender : ''}
+              t={touched.gender ? errors.gender : ''}
               error={touched.gender && Boolean(errors.gender)}
               onChange={handleChange}
               select
@@ -189,11 +189,11 @@ const EditProfile: React.FC<EditProfileProps> = ({ header, currentUser, currentP
               id="telephone"
               label="Telephone"
               fullWidth
-              margin="normal"
+              //margin="normal"
               name="telephone"
               placeholder="Your phone"
               autoComplete="telephone"
-              helperText={touched.telephone ? errors.telephone : ''}
+              t={touched.telephone ? errors.telephone : ''}
               error={touched.telephone && Boolean(errors.telephone)}
               value={values.telephone}
               onChange={handleChange}
@@ -203,11 +203,11 @@ const EditProfile: React.FC<EditProfileProps> = ({ header, currentUser, currentP
               id="address"
               label="Address"
               fullWidth
-              margin="normal"
+              //margin="normal"
               name="address"
               placeholder="Where you live"
               autoComplete="address"
-              helperText={touched.address ? errors.address : ''}
+              t={touched.address ? errors.address : ''}
               error={touched.address && Boolean(errors.address)}
               value={values.address}
               onChange={handleChange}
@@ -217,11 +217,11 @@ const EditProfile: React.FC<EditProfileProps> = ({ header, currentUser, currentP
               id="description"
               label="Description"
               fullWidth
-              margin="normal"
+              //margin="normal"
               name="description"
               placeholder="Tell us about yourself"
               autoComplete="description"
-              helperText={touched.description ? errors.description : ''}
+              t={touched.description ? errors.description : ''}
               error={touched.description && Boolean(errors.description)}
               value={values.description}
               onChange={handleChange}
